refactor: apply dependency injection and update module dependencies
